LayoutTests/http/tests/inspector/network/set-resource-caching-disabled-memory-cache-expected.txt

Test for `Network.setResourceCachingDisabled` for memory cache loads.


== Running test suite: Network.SetResourceCachingDisabled.MemoryCache
-- Running test setup.
-- Running test case: Network.SetResourceCachingDisabled.MemoryCache
PASS: Resource should be exist.
PASS: statusCode should be 200
PASS: responseSource should be Symbol(network)
